@@ -39,6 +39,7 @@
         "@types/styled-components": "^5.1.7",
         "@typescript-eslint/eslint-plugin": "^4.13.0",
         "@typescript-eslint/parser": "^4.13.0",
+        "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
         "babel-core": "^6.26.3",
         "babel-jest": "^25.3.0",
         "babel-loader": "^8.1.0",
@@ -47,7 +48,6 @@
         "css-loader": "^5.2.7",
         "cypress": "^8.3.1",
         "enzyme": "^3.11.0",
-        "enzyme-adapter-react-16": "^1.15.2",
         "eslint": "^6.8.0",
         "eslint-config-airbnb": "^18.1.0",
         "eslint-config-prettier": "^6.10.1",
@@ -93,13 +93,13 @@
     ],
     "dependencies": {
         "firebase": "^8.2.0",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
+        "react": "^17.0.0",
+        "react-dom": "^17.0.0",
         "react-intl": "^4.4.0",
         "react-query": "^3.34.0",
         "react-rainbow-components": "1.29.0-canary.c10cd9f",
         "react-redux": "^7.2.0",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^5.0.0",
         "redux": "^4.0.5",
         "redux-form": "^8.3.5",
         "redux-thunk": "^2.3.0",
