@@ -1,14 +1,15 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Route, Redirect } from 'react-router-dom';
+import { Route, Redirect, useLocation, useNavigate } from 'react-router-dom';
 import { useAuthState } from '@rainbow-modules/firebase-hooks';
+import { isRouterV6 } from '@rainbow-modules/app';
 import getRedirectTo from '../../helpers/getRedirectTo';
 
 const Private = (props) => {
     // eslint-disable-next-line react/prop-types
-    const { location, component, redirect, children, ...rest } = props;
+    const { location, component: Component, redirect, children, ...rest } = props;
     const isAuth = useAuthState();
-    const Component = component;
+
     if (isAuth) {
         if (Component) {
             // eslint-disable-next-line react/jsx-props-no-spreading
@@ -29,7 +30,31 @@ const Private = (props) => {
     return null;
 };
 
-const WhenAuthenticated = (props) => {
+const WhenAuthenticatedV6 = (props) => {
+    // eslint-disable-next-line react/prop-types
+    const { component: Component, redirect, children, ...rest } = props;
+    const location = useLocation();
+    const navigate = useNavigate();
+    const isAuth = useAuthState();
+
+    useEffect(() => {
+        if (isAuth === false) {
+            navigate(redirect);
+        }
+    }, [isAuth, navigate, redirect]);
+
+    if (isAuth) {
+        if (Component) {
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            return <Component location={location} {...rest} />;
+        }
+        return <div>{children}</div>;
+    }
+    return null;
+};
+
+const WhenAuthenticatedV5 = (props) => {
+    // eslint-disable-next-line react/prop-types
     const { path, redirect, component, children } = props;
 
     return (
@@ -44,6 +69,8 @@ const WhenAuthenticated = (props) => {
         />
     );
 };
+
+const WhenAuthenticated = isRouterV6 ? WhenAuthenticatedV6 : WhenAuthenticatedV5;
 
 WhenAuthenticated.propTypes = {
     /** The route path the component will used to match against the browser URL. */
