@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Route, Redirect } from 'react-router-dom';
+import { Route, Redirect, useLocation, useNavigate } from 'react-router-dom';
 import { useAuthState } from '@rainbow-modules/firebase-hooks';
+import { isRouterV6 } from '@rainbow-modules/app';
 import getRedirectTo from '../../helpers/getRedirectTo';
 
 const Private = (props) => {
@@ -29,7 +30,32 @@ const Private = (props) => {
     return null;
 };
 
-const WhenNoAuthenticated = (props) => {
+const WhenNoAuthenticatedV6 = (props) => {
+    // eslint-disable-next-line react/prop-types
+    const { component, redirect, children, ...rest } = props;
+    const location = useLocation();
+    const navigate = useNavigate();
+    const isAuth = useAuthState();
+    const Component = component;
+
+    useEffect(() => {
+        if (isAuth === true) {
+            navigate(redirect);
+        }
+    }, [isAuth, navigate, redirect]);
+
+    if (isAuth === false) {
+        if (Component) {
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            return <Component location={location} {...rest} />;
+        }
+        return children;
+    }
+    return null;
+};
+
+const WhenNoAuthenticatedV5 = (props) => {
+    // eslint-disable-next-line react/prop-types
     const { path, redirect, component, children } = props;
 
     return (
@@ -44,6 +70,8 @@ const WhenNoAuthenticated = (props) => {
         />
     );
 };
+
+const WhenNoAuthenticated = isRouterV6 ? WhenNoAuthenticatedV6 : WhenNoAuthenticatedV5;
 
 WhenNoAuthenticated.propTypes = {
     path: PropTypes.string,
